Checks if user does exist before following
The API will is able to check if the user exists, if not then it will use the if statement and injecting a SQL connection query to check the user on the database. 

Also checking all user details and who they are following and their followers. This is able to check all user details and along with their following and follower details.